Implement letterSpacing option on Android.
Because Android TextView#setLetterSpacing expects the letter spacing
value in 'EM' unit, where one is defined by the width of the letter 'M'.
Typical values for slight expansion will be around 0.05. Negative values
will tighten text. So we use NaN as not defined value.

This method calculates an 'EM' *approach* based on the input and the
calculated font size. With common fonts this calculated value renders a
similar spacing and layout to iOS.